Complete chapter 2.6!
We now parse prefix and infix expressions that consist of identifiers
and integer literals. As per the book, this is done with some
Pratt-esque parsing, and several test examples. I've had to make some
rust-y changes to the suggestions in the book, obviously, but the
implementation is pretty straightforward.

We can now parse expressions such as:

```
3 + 5 * 10 / somevar == anothervar - 20;
```